requirements.txt
Django==6.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.1
python-dotenv==1.0.0
